Version 3.31.32 (based on 9e02e98771dcb00ed7332eaeeb78464a8b87afe5)
Performance and stability improvements on all platforms.

Cr-Commit-Position: refs/heads/candidates@{#25306}

+++ src/heap/gc-idle-time-handler.cc
@@ -191,7 +191,7 @@ bool GCIdleTimeHandler::ShouldDoFinalIncrementalMarkCompact(
 // that this currently may trigger a full garbage collection.
 GCIdleTimeAction GCIdleTimeHandler::Compute(double idle_time_in_ms,
                                             HeapState heap_state) {
-  if (static_cast<int>(idle_time_in_ms) <= 0) {
+  if (idle_time_in_ms <= 0.0) {
     if (heap_state.incremental_marking_stopped) {
       if (ShouldDoContextDisposalMarkCompact(
               heap_state.contexts_disposed,
